backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pillow>=10.0.0
numpy>=1.24.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
databases>=0.7.0
aiofiles>=23.1.0
python-magic-bin; platform_system == "Windows"
# ---- AquaSafe AI (Added Dependencies) ----
ultralytics>=8.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
reportlab>=4.0.0
